cambio a neonDb

+++ app/models/model_cliente.py
@@ -1,5 +1,4 @@
 from sqlalchemy import Column, Integer, String, Text, DateTime
-from sqlalchemy.dialects.mysql import TINYINT
 from sqlalchemy.sql import func
 
 from app.database import Base
@@ -13,7 +12,7 @@ class Cliente(Base):
     telefono = Column(String(20))
     email = Column(String(100))
     direccion = Column(Text)
-    nit = Column(String(20))  
-    tipo_cliente = Column(String(50))  
+    nit = Column(String(20)) 
+    tipo_cliente = Column(String(50))  # Persona, Empresa, Escuela, etc.
     fecha_registro = Column(DateTime, server_default=func.now())
-    activo = Column(TINYINT, default=1)  
+    activo = Column(Integer, default=1)
