+++ app/models/model_inventario.py
@@ -1,19 +1,15 @@
 from sqlalchemy import Column, Integer, String, Text, DateTime, DECIMAL, ForeignKey
-from sqlalchemy.dialects.mysql import TINYINT
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
-
 from app.database import Base
-
 
 class TipoMovimientoInventario(Base):
     __tablename__ = "tipo_movimiento_inventario"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    nombre = Column(String(50))  
+    nombre = Column(String(50)) 
     descripcion = Column(Text)
-    signo = Column(Integer)  
-
+    signo = Column(Integer) 
 
 class MovimientoInventario(Base):
     __tablename__ = "movimiento_inventario"
@@ -22,10 +18,12 @@ class MovimientoInventario(Base):
     fecha = Column(DateTime, server_default=func.now())
     id_usuario = Column(Integer, ForeignKey("usuario.id"))
     id_tipo_movimiento = Column(Integer, ForeignKey("tipo_movimiento_inventario.id"), nullable=False)
+
     id_ubicacion_origen = Column(Integer, ForeignKey("ubicacion.id"))
     id_sububicacion_origen = Column(Integer, ForeignKey("sububicacion.id"))
     id_ubicacion_destino = Column(Integer, ForeignKey("ubicacion.id"))
     id_sububicacion_destino = Column(Integer, ForeignKey("sububicacion.id"))
+
     id_referencia = Column(Integer)  # id de la compra/venta/traslado que originó esto
     tabla_referencia = Column(String(20))  # 'compra', 'venta', 'traslado'
     referencia = Column(String(100))
@@ -35,7 +33,6 @@ class MovimientoInventario(Base):
     detalles = relationship(
         "MovimientoInventarioDetalle", back_populates="movimiento", cascade="all, delete-orphan"
     )
-
 
 class MovimientoInventarioDetalle(Base):
     __tablename__ = "movimiento_inventario_detalle"
@@ -50,7 +47,6 @@ class MovimientoInventarioDetalle(Base):
     movimiento = relationship("MovimientoInventario", back_populates="detalles")
     producto = relationship("Producto")
 
-
 class InventarioFisico(Base):
     __tablename__ = "inventario_fisico"
 
@@ -59,15 +55,14 @@ class InventarioFisico(Base):
     id_ubicacion = Column(Integer, ForeignKey("ubicacion.id"))
     id_sububicacion = Column(Integer, ForeignKey("sububicacion.id"))
     fecha = Column(DateTime, server_default=func.now())
-    stock_sistema = Column(DECIMAL(12, 2)) 
-    stock_real = Column(DECIMAL(12, 2)) 
-    diferencia = Column(DECIMAL(12, 2))  
+    stock_sistema = Column(DECIMAL(12, 2))  
+    stock_real = Column(DECIMAL(12, 2))  
+    diferencia = Column(DECIMAL(12, 2)) 
     id_usuario = Column(Integer, ForeignKey("usuario.id"))
     observaciones = Column(Text)
-    ajustado = Column(TINYINT, default=0) 
+    ajustado = Column(Integer, default=0)
 
     producto = relationship("Producto")
-
 
 class TrasladoSucursal(Base):
     __tablename__ = "traslado_sucursal"
@@ -76,16 +71,18 @@ class TrasladoSucursal(Base):
     fecha = Column(DateTime, server_default=func.now())
     id_producto = Column(Integer, ForeignKey("producto.id"), nullable=False)
     cantidad = Column(DECIMAL(12, 2), nullable=False)
+
     id_ubicacion_origen = Column(Integer, ForeignKey("ubicacion.id"))
     id_ubicacion_destino = Column(Integer, ForeignKey("ubicacion.id"))
     id_sububicacion_origen = Column(Integer, ForeignKey("sububicacion.id"))
     id_sububicacion_destino = Column(Integer, ForeignKey("sububicacion.id"))
-    costo_unitario = Column(DECIMAL(12, 2))  # sale a precio de costo, gana la sucursal destino
-    metodo_traslado = Column(String(50))  # "Uber Moto" (urgente) o "Empleado Interno" (no urgente)
+
+    costo_unitario = Column(DECIMAL(12, 2))  
+    metodo_traslado = Column(String(50)) 
     id_usuario_autoriza = Column(Integer, ForeignKey("usuario.id"))
     id_usuario_recibe = Column(Integer, ForeignKey("usuario.id"))
     fecha_recepcion = Column(DateTime)
-    estado = Column(String(20), default="EnProceso")  # EnProceso, Recibido, Completado
+    estado = Column(String(20), default="EnProceso")  
     observaciones = Column(Text)
 
     producto = relationship("Producto")
@@ -96,11 +93,11 @@ class Alerta(Base):
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     fecha = Column(DateTime, server_default=func.now())
-    tipo = Column(String(30))  # stock_bajo, stock_excedido, sin_movimiento, vencimiento
+    tipo = Column(String(30)) 
     mensaje = Column(Text)
     id_producto = Column(Integer, ForeignKey("producto.id"))
     id_usuario_destino = Column(Integer, ForeignKey("usuario.id"))
-    leida = Column(TINYINT, default=0)
+    leida = Column(Integer, default=0)
     fecha_lectura = Column(DateTime)
 
     producto = relationship("Producto")
